upgrade lexicon-library & spring + do some refactoring
