+++ src/main/java/com/emvista/emailreplyparser/model/FragmentDTO.java
@@ -1,4 +1,4 @@
-package com.emvista.edlio.emailreplyparser;
+package com.emvista.emailreplyparser.model;
 
 import java.util.List;
