--- src/test/resources/application.properties
@@ -1,3 +1,0 @@
-#lexicon lib
-com.emvista.lexicon.db.database=Lex_Fr
-com.emvista.lexicon.db.uri=mongodb://venus.em:27017
